Feature/impl service

+++ src/main/java/ru/skypro/homework/service/impl/AdService.java
@@ -1,27 +1,48 @@
 package ru.skypro.homework.service.impl;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
+import org.springframework.web.multipart.MultipartFile;
 import ru.skypro.homework.dto.ad.Ad;
 import ru.skypro.homework.dto.ad.Ads;
 import ru.skypro.homework.dto.ad.CreateOrUpdateAd;
 import ru.skypro.homework.dto.ad.ExtendedAd;
 import ru.skypro.homework.mapper.AdsMapper;
 import ru.skypro.homework.models.AdEntity;
+import ru.skypro.homework.models.UserEntity;
 import ru.skypro.homework.repository.AdRepository;
+import ru.skypro.homework.repository.UserRepository;
 
+import java.io.*;
+import java.nio.file.Files;
+import java.nio.file.Path;
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
+
+
 
 @Service
 public class AdService {
 
-    @Autowired
-    private AdRepository adRepository;
+    private final AdRepository adRepository;
 
-    @Autowired
-    private AdsMapper adsMapper;
+    private final AdsMapper adsMapper;
+    @Value("${avatars.ads.dir}")
+    private String adsDir;
+
+    private final FileService fileService;
+    private final UserRepository userRepository;
+
+    public AdService(AdRepository adRepository, AdsMapper adsMapper, FileService fileService, UserRepository userRepository) {
+        this.adRepository = adRepository;
+        this.adsMapper = adsMapper;
+        this.fileService = fileService;
+        this.userRepository = userRepository;
+    }
 
     public Ads getAllAds() {
         List<AdEntity> allAds = adRepository.findAll();
@@ -35,9 +56,14 @@ public class AdService {
         return adsMapper.toExtendedAd(user);
     }
 
-    public void deleteAdById(long id) {
-        adRepository.deleteById(id);
-        adRepository.existsById(id);
+    public boolean deleteAdById(long id) {
+        boolean exists = adRepository.existsById(id);
+        if (exists) {
+            adRepository.deleteById(id);
+            return true;
+        }
+        return false;
+
     }
 
     public Ads getAdsAuthorizedUser() {
@@ -63,4 +89,36 @@ public class AdService {
         return newAd;
     }
 
+    public void updateImage(long idAd, MultipartFile image) throws IOException {
+        AdEntity ad = adRepository.findById(idAd).get();
+        String path = ad + "." + getExtensions((Objects.requireNonNull(image.getOriginalFilename())));
+        String filePath = fileService.uploadFile(adsDir, path, image);
+        ad.setImage(filePath);
+        adRepository.save(ad);
+    }
+
+    private String getExtensions(String fileName) {
+        return fileName.substring(fileName.lastIndexOf(".") + 1);
+    }
+
+    public Ad addAd(CreateOrUpdateAd ad, MultipartFile image) throws IOException {
+        String name = SecurityContextHolder.getContext().getAuthentication().getName();
+        UserEntity user = userRepository.findByEmail(name).orElseThrow(() ->
+                new UsernameNotFoundException("Пользователь не найден: " + name));
+        String path = ad + "." + getExtensions((Objects.requireNonNull(image.getOriginalFilename())));
+        String filePath = fileService.uploadFile(adsDir, path, image);
+        AdEntity adEntity = adsMapper.toAdEntity(ad, user, filePath);
+        adRepository.save(adEntity);
+        return adsMapper.toAd(adEntity);
+    }
+
+
+    public boolean getIsAdOwner(Long id) {
+        String name = SecurityContextHolder.getContext().getAuthentication().getName();
+        UserEntity user = userRepository.findByEmail(name).orElseThrow(() ->
+                new UsernameNotFoundException("Пользователь не найден: " + name));
+        AdEntity adEntity = adRepository.findById(id).get();
+        return user.getId().equals(adEntity.getAuthor().getId());
+
+    }
 }
